backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
apscheduler==3.10.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
